Featureback/board

See merge request s03-final/s03p31a205!79

+++ django_server/photory/board/views.py
@@ -30,33 +30,30 @@ def article_create(request):
         serializer.save()
         return Response(serializer.data)
 
-@api_view(['PUT'])
-@permission_classes([IsAuthenticated])
-def article_update(request,article_pk):
-    article =Article.objects.get(pk=article_pk)
-    article.bookcover = request.data['bookcover']
-    article.title = request.data['title']
-    article.writer = request.data['writer']
-    article.content = request.data['content']
-    article.category = request.data['category']
-    article.save()
-    return Response(status =200)
-
-
-@api_view(['DELETE'])
+@api_view(['DELETE','PUT'])
 @permission_classes([IsAuthenticated]) 
-def article_delete(request, article_pk):
-	article = Article.objects.get(pk = article_pk)
-	article.delete()
-	return Response(status=200)
+def article_delete_update(request, article_pk): 
+    if request.method == 'DELETE':
+        article = Article.objects.get(pk = article_pk)
+        article.delete()
+        return Response(status=200)
+    else:
+        article =Article.objects.get(pk=article_pk)
+        article.bookcover = request.data['bookcover']
+        article.title = request.data['title']
+        article.writer = request.data['writer']
+        article.content = request.data['content']
+        article.category = request.data['category']
+        article.save()
+        return Response(status =200)
 
 def index(request, pagenum,pagenum2):
     articles = Article.objects.order_by('-pk')
     if pagenum2 ==0:
         pagenum2=1
     pages = [i for i in range(1,len(articles)+1)]
-    page2 = max(1,min((pages[-1]-1//10,pagenum2)))
-    pages= pages[(page2-1)*10:(page2)*10]
+    page2 = max(1,min((pages[-1]-1//8,pagenum2)))
+    pages= pages[(page2-1)*8:(page2)*8]
     if pagenum<pages[0]:
         pagenum=pages[0]
     elif pagenum>pages[-1]:
